requirements.txt
Flask==2.3.3
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
pyodbc==5.1.0
gunicorn==22.0.0